refactor cancel into service
