+++ src/test/javascript/spec/component/orion/orion-exercise-assessment-dashboard.component.spec.ts
@@ -8,7 +8,7 @@ import { ExerciseType } from 'app/entities/exercise.model';
 import { TutorParticipationStatus } from 'app/entities/participation/tutor-participation.model';
 import { ProgrammingExercise } from 'app/entities/programming-exercise.model';
 import { OrionConnectorService } from 'app/shared/orion/orion-connector.service';
-import { BehaviorSubject, Observable, of, throwError } from 'rxjs';
+import { BehaviorSubject, of, throwError } from 'rxjs';
 import { MockComponent, MockPipe, MockProvider } from 'ng-mocks';
 import { HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { JhiAlertService } from 'ng-jhipster';
@@ -21,7 +21,6 @@ import { ExerciseService } from 'app/exercises/shared/exercise/exercise.service'
 import { ActivatedRoute, convertToParamMap } from '@angular/router';
 import { OrionAssessmentService } from 'app/orion/assessment/orion-assessment.service';
 import { OrionButtonComponent } from 'app/shared/orion/orion-button/orion-button.component';
-import { ProgrammingAssessmentManualResultService } from 'app/exercises/programming/assess/manual-result/programming-assessment-manual-result.service';
 
 chai.use(sinonChai);
 const expect = chai.expect;
@@ -50,7 +49,6 @@ describe('OrionExerciseAssessmentDashboardComponent', () => {
             providers: [
                 MockProvider(OrionConnectorService),
                 MockProvider(OrionAssessmentService),
-                MockProvider(ProgrammingAssessmentManualResultService),
                 MockProvider(ExerciseService),
                 MockProvider(JhiAlertService),
                 MockProvider(TranslateService),
@@ -127,33 +125,5 @@ describe('OrionExerciseAssessmentDashboardComponent', () => {
         expect(getForTutorsStub).to.have.been.calledOnceWithExactly(10);
         expect(orionStateStub).to.have.been.calledOnceWithExactly();
         expect(errorSpy).to.have.been.calledOnceWithExactly('error.http.400');
-    }));
-    it('should cancel lock correctly', fakeAsync(() => {
-        const orionState = { opened: 40, building: false, cloning: true } as OrionState;
-        const stateObservable = new BehaviorSubject(orionState);
-        const orionStateStub = stub(orionConnectorService, 'state');
-        orionStateStub.returns(stateObservable);
-
-        const cancelStub = stub(TestBed.inject(ProgrammingAssessmentManualResultService), 'cancelAssessment');
-        cancelStub.returns(new Observable());
-
-        const response = of(new HttpResponse({ body: programmingExercise, status: 200 }));
-        const getForTutorsStub = stub(TestBed.inject(ExerciseService), 'getForTutors');
-        getForTutorsStub.returns(response);
-
-        comp.ngOnInit();
-        tick();
-
-        expect(comp.exerciseId).to.be.equals(10);
-        expect(comp.exercise).to.be.deep.equals(programmingExercise);
-        expect(comp.orionState).to.be.deep.equals(orionState);
-
-        // @ts-ignore
-        comp.setActiveSubmissionId(24);
-
-        stateObservable.next({ ...orionState, cloning: false });
-        tick();
-
-        expect(cancelStub).to.have.been.calledOnceWithExactly(24);
     }));
 });
